refactor(hmr): updated as per review comments
