+++ packages/holocron-dev-server/src/components/Document.jsx
