+++ packages/holocron-dev-server/src/components/HolocronHmrWrapper.jsx
