+++ packages/holocron-dev-server/src/components/RegisterModule.jsx
