+++ packages/holocron-dev-server/src/middleware/ParrotScenarios.js
