+++ packages/holocron-dev-server/src/middleware/html.js
