+++ packages/holocron-dev-server/src/middleware/index.js
