+++ packages/holocron-dev-server/src/middleware/request-accepted.js
