+++ packages/holocron-dev-server/src/utils/config.js
