+++ packages/holocron-dev-server/src/utils/module-map.js
