+++ packages/holocron-dev-server/src/utils/paths.js
