+++ packages/holocron-dev-server/src/utils/rendering.js
