+++ packages/holocron-dev-server/src/utils/statics.js
