+++ packages/holocron-dev-server/src/utils/virtual-file-system.js
