+++ packages/holocron-dev-server/src/utils/watcher.js
