+++ packages/holocron-dev-server/src/webpack/configs/bundles.js
