+++ packages/holocron-dev-server/src/webpack/configs/fragments.js
