+++ packages/holocron-dev-server/src/webpack/configs/index.js
