+++ packages/holocron-dev-server/src/webpack/configs/loaders.js
