+++ packages/holocron-dev-server/src/webpack/plugins/holocron-webpack-plugin.js
