+++ packages/holocron-dev-server/src/webpack/plugins/index.js
